# задание/openapi.yml
openapi: "3.0.1"
info:
  title: Tender Management API
  version: "1.0"
  description: |
    API для управления тендерами и предложениями. 

    Основные функции API включают управление тендерами (создание, изменение, получение списка) и управление предложениями (создание, изменение, получение списка).
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер API
paths:
...
  /tenders/new:
    post:
      summary: Создание нового тендера
      description: Создание нового тендера с заданными параметрами.
      operationId: createTender
      requestBody:
        description: Данные нового тендера.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/tenderName"
                description:
                  $ref: "#/components/schemas/tenderDescription"
                serviceType:
                  $ref: "#/components/schemas/tenderServiceType"
                organizationId:
                  $ref: "#/components/schemas/organizationId"
                creatorUsername:
                  $ref: "#/components/schemas/username"
              required:
                - name
                - description
                - serviceType
                - organizationId
                - creatorUsername
      responses:
        "200":
          description: Тендер успешно создан. Сервер присваивает уникальный идентификатор и время создания.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tender"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "403":
          description: Недостаточно прав для выполнения действия.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
...
  /bids/new:
    post:
      summary: Создание нового предложения
      description: Создание предложения для существующего тендера.
      operationId: createBid
      requestBody:
        description: Данные нового предложения.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/bidName"
                description:
                  $ref: "#/components/schemas/bidDescription"
                tenderId:
                  $ref: "#/components/schemas/tenderId"
                authorType:
                  $ref: "#/components/schemas/bidAuthorType"
                authorId:
                  $ref: "#/components/schemas/bidAuthorId"
              required:
                - name
                - description
                - tenderId
                - authorType
                - authorId
      responses:
        "200":
          description: Предложение успешно создано. Сервер присваивает уникальный идентификатор и время создания.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bid"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "403":
          description: Недостаточно прав для выполнения действия.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "404":
          description: Тендер не найден.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
  /bids/my:
    get:
      summary: Получение списка ваших предложений
      description: |
        Получение списка предложений текущего пользователя.

        Для удобства использования включена поддержка пагинации.
      operationId: getUserBids
      parameters:
        - $ref: "#/components/parameters/paginationLimit"
        - $ref: "#/components/parameters/paginationOffset"
        - name: username
          in: query
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Список предложений пользователя, отсортированный по алфавиту.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/bid"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
...
  /bids/{bidId}/status:
    get:
      summary: Получение текущего статуса предложения
      description: Получить статус предложения по его уникальному идентификатору.
      operationId: getBidStatus
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/bidId"
        - name: username
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Текущий статус предложения.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bidStatus"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "403":
          description: Недостаточно прав для выполнения действия.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "404":
          description: Предложение не найдено.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
    put:
      summary: Изменение статуса предложения
      description: Изменить статус предложения по его уникальному идентификатору.
      operationId: updateBidStatus
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/bidId"
        - name: status
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/bidStatus"
        - name: username
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Статус предложения успешно изменен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bid"
        "400":
          description: Неверный формат запроса или его параметры.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "403":
          description: Недостаточно прав для выполнения действия.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "404":
          description: Предложение не найдено.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
...
components:
  schemas:
    username:
      type: string
      description: Уникальный slug пользователя.
      example: test_user
    tenderStatus:
      type: string
      description: Статус тендер
      enum:
        - Created
        - Published
        - Closed
    tenderServiceType:
      type: string
      description: Вид услуги, к которой относиться тендер
      enum:
        - Construction
        - Delivery
        - Manufacture
    tenderId:
      type: string
      description: Уникальный идентификатор тендера, присвоенный сервером.
      example: 550e8400-e29b-41d4-a716-446655440000
      maxLength: 100
    tenderName:
      type: string
      description: Полное название тендера
      maxLength: 100
    tenderDescription:
      type: string
      description: Описание тендера
      maxLength: 500
    tenderVersion:
      type: integer
      description: Номер версии посел правок
      format: int32
      minimum: 1
      default: 1
    organizationId:
      type: string
      description: Уникальный идентификатор организации, присвоенный сервером.
      example: 550e8400-e29b-41d4-a716-446655440000
      maxLength: 100
    tender:
      type: object
      description: Информация о тендере
      properties:
        id:
          $ref: "#/components/schemas/tenderId"
        name:
          $ref: "#/components/schemas/tenderName"
        description:
          $ref: "#/components/schemas/tenderDescription"
        serviceType:
          $ref: "#/components/schemas/tenderServiceType"
        status:
          $ref: "#/components/schemas/tenderStatus"
        organizationId:
          $ref: "#/components/schemas/organizationId"
        version:
          $ref: "#/components/schemas/tenderVersion"
        createdAt:
          type: string
          description: |
            Серверная дата и время в момент, когда пользователь отправил тендер на создание.
            Передается в формате RFC3339.
          example: 2006-01-02T15:04:05Z07:00
        
      required:
        - id
        - name
        - description
        - serviceType
        - status
        - organizationId
        - version
        - createdAt
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
        name: Доставка товары Казань - Москва
        description: Нужно доставить оборудовоние для олимпиады по робототехники
        status: Created
        serviceType: Delivery
        version: 1
        createdAt: 2006-01-02T15:04:05Z07:00
    bidStatus:
      type: string
      description: Статус предложения
      enum:
        - Created
        - Published
        - Canceled
...
    bidId:
      type: string
      description: Уникальный идентификатор предложения, присвоенный сервером.
      example: 550e8400-e29b-41d4-a716-446655440000
      maxLength: 100
    bidName:
      type: string
      description: Полное название предложения
      maxLength: 100
    bidDescription:
      type: string
      description: Описание предложения
      maxLength: 500
...
    bidAuthorType:
      type: string
      description: Тип автора
      enum:
        - Organization
        - User
    bidAuthorId:
      type: string
      description: Уникальный идентификатор автора предложения, присвоенный сервером.
      example: 550e8400-e29b-41d4-a716-446655440000
      maxLength: 100
    bidVersion:
      type: integer
      description: Номер версии посел правок
      format: int32
      minimum: 1
      default: 1
...
    bid:
      type: object
      description: Информация о предложении
      properties:
        id:
          $ref: "#/components/schemas/bidId"
        name:
          $ref: "#/components/schemas/bidName"
        description:
          $ref: "#/components/schemas/bidDescription"
        status:
          $ref: "#/components/schemas/bidStatus"
        tenderId:
          $ref: "#/components/schemas/tenderId"
        authorType:
          $ref: "#/components/schemas/bidAuthorType"
        authorId:
          $ref: "#/components/schemas/bidAuthorId"
        version:
          $ref: "#/components/schemas/bidVersion"
        createdAt:
          type: string
          description: |
            Серверная дата и время в момент, когда пользователь отправил предложение на создание.
            Передается в формате RFC3339.
          example: 2006-01-02T15:04:05Z07:00
        
      required:
        - id
        - name
        - description
        - status
        - tenderId
        - createdAt
        - authorType
        - authorId
        - version
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
        name: Доставка товаров Алексей
        status: Created
        authorType: User
        authorId: 61a485f0-e29b-41d4-a716-446655440000
        version: 1
        createdAt: 2006-01-02T15:04:05Z07:00
    errorResponse:
      type: object
      description: Используется для возвращения ошибки пользователю
      properties:
        reason:
          type: string
          description: Описание ошибки в свободной форме
          minLength: 5
      required:
        - reason
      example:
        reason: <объяснение, почему запрос пользователя не может быть обработан>
  parameters:
    paginationLimit:
      in: query
      name: limit
      required: false
      description: |
        Максимальное число возвращаемых объектов. Используется для запросов с пагинацией.

        Сервер должен возвращать максимальное допустимое число объектов.
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 50
        default: 5
    paginationOffset:
      in: query
      name: offset
      required: false
      description: |
        Какое количество объектов должно быть пропущено с начала. Используется для запросов с пагинацией.
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0